Remove option to undo deleting exercise
This is a lazy fix for a bug where sets werent shown anymore after un-deleting the ExerciseHolder. The reason the undo functionality is not necessary anymore is because in the future, exercises should be deleted by deleting the last set, so there wont be important data loss for the user anyways. Maybe this feature will be implemented later, but for now it's not necessary.